#6 reformatted, impement parts of worker
